+++ ruby/lib/gitaly_server.rb
@@ -6,6 +6,7 @@ require_relative 'gitaly_server/utils.rb'
 require_relative 'gitaly_server/commit_service.rb'
 require_relative 'gitaly_server/diff_service.rb'
 require_relative 'gitaly_server/ref_service.rb'
+require_relative 'gitaly_server/operations_service.rb'
 
 module GitalyServer
   REPO_PATH_HEADER = 'gitaly-repo-path'.freeze
@@ -18,5 +19,6 @@ module GitalyServer
     server.handle(CommitService.new)
     server.handle(DiffService.new)
     server.handle(RefService.new)
+    server.handle(OperationsService.new)
   end
 end